chores(gateway): move provider factory into a separated module
