+++ apps/gateway/src/services/model-resolver.ts
@@ -1,28 +1,14 @@
-import { createAmazonBedrock } from "@ai-sdk/amazon-bedrock";
-import { createVertex } from "@ai-sdk/google-vertex";
-import { createGroq } from "@ai-sdk/groq";
-import { fromTemporaryCredentials } from "@aws-sdk/credential-providers";
-import { withCanonicalIdsForBedrock } from "@hebo-ai/gateway/providers/bedrock";
-import { withCanonicalIdsForGroq } from "@hebo-ai/gateway/providers/groq";
-import { withCanonicalIdsForVertex } from "@hebo-ai/gateway/providers/vertex";
-import { withCanonicalIdsForVoyage } from "@hebo-ai/gateway/providers/voyage";
 import QuickLRU from "quick-lru";
-import { createVoyage } from "voyage-ai-provider";
 
 import { getSecret } from "@hebo/shared-api/utils/secrets";
 
 import type { createDbClient } from "~api/lib/db/client";
 
 export type DbClient = ReturnType<typeof createDbClient>;
-import type {
-  ApiKeyProviderConfig,
-  BedrockProviderConfig,
-  Models,
-  ProviderSlug,
-  VertexProviderConfig,
-} from "~api/modules/providers/types";
+import type { Models, ProviderSlug } from "~api/modules/providers/types";
 
-import { buildWifOptions, injectMetadataCredentials } from "./aws-wif";
+import { injectMetadataCredentials } from "./aws-wif";
+import { createProvider } from "./provider-factory";
 
 import type { ProviderV3 } from "@ai-sdk/provider";
 import type { ProviderRegistry } from "@hebo-ai/gateway";
@@ -60,51 +46,6 @@ export async function loadProviderSecrets() {
     vertexLocation,
     vertexProject,
   };
-}
-
-export function createProvider(
-  slug: ProviderSlug,
-  config: unknown,
-): ProviderV3 {
-  switch (slug) {
-    case "bedrock": {
-      const { bedrockRoleArn, region } = config as BedrockProviderConfig;
-      return withCanonicalIdsForBedrock(
-        createAmazonBedrock({
-          region,
-          credentialProvider: fromTemporaryCredentials({
-            params: { RoleArn: bedrockRoleArn },
-            clientConfig: region ? { region } : undefined,
-          }),
-        }),
-      );
-    }
-    case "groq": {
-      const { apiKey } = config as ApiKeyProviderConfig;
-      return withCanonicalIdsForGroq(createGroq({ apiKey }));
-    }
-    case "vertex": {
-      const { serviceAccountEmail, audience, location, project } =
-        config as VertexProviderConfig;
-      return withCanonicalIdsForVertex(
-        createVertex({
-          googleAuthOptions: {
-            credentials: buildWifOptions(audience, serviceAccountEmail),
-            scopes: ["https://www.googleapis.com/auth/cloud-platform"],
-          },
-          location,
-          project,
-        }),
-      );
-    }
-    case "voyage": {
-      const { apiKey } = config as ApiKeyProviderConfig;
-      return withCanonicalIdsForVoyage(createVoyage({ apiKey }));
-    }
-    default: {
-      throw new Error(`Unsupported provider: ${slug}`);
-    }
-  }
 }
 
 export async function resolveModelId(
